# server/docs/swagger.yaml
basePath: /
definitions:
  server.CreateGameRequest:
    properties:
      gameVariant:
        type: string
      playerNum:
        type: integer
    type: object
  server.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  server.GameResponse:
    properties:
      currentPlayers:
        type: integer
      id:
        type: integer
      maxPlayers:
        type: integer
      variant:
        type: string
    type: object
  server.JoinGameRequest:
    properties:
      username:
        type: string
    type: object
  server.JoinGameResponse:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  server.Response:
    properties:
      message:
        type: string
    type: object
  server.SaveGameRequest:
    properties:
      name:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: rvrelay@gmail.com
    name: Project Documentation
    url: https://github.com/rafisto/chinese-checkers
  description: This is the API for the Chinese Checkers game. It allows you to create
    and join games, and play the game with other players.
  termsOfService: http://swagger.io/terms/
  title: Chinese Checkers API
  version: 1.0-sprint1
paths:
  /games:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Successfully received all active games
          schema:
            items:
              $ref: '#/definitions/server.GameResponse'
            type: array
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get all currently active games
      tags:
      - Game
    post:
      consumes:
      - application/json
      parameters:
      - description: Initial game parameters
        in: body
        name: playerNum
        required: true
        schema:
          $ref: '#/definitions/server.CreateGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created game
          schema:
            $ref: '#/definitions/server.Response'
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new game provided its initial parameters
      tags:
      - Game
  /games/{game_id}/bot:
    post:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Successfully added the bot
          schema:
            $ref: '#/definitions/server.Response'
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Add a bot to the game
      tags:
      - Game
  /games/{game_id}/join:
    post:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      - description: Player username
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/server.JoinGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully joined the game
          schema:
            $ref: '#/definitions/server.JoinGameResponse'
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Join a game by its ID, provided the username
      tags:
      - Game
  /games/{game_id}/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: game_id
        required: true
        type: string
      - description: Game name
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/server.SaveGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully saved the game
          schema:
            $ref: '#/definitions/server.Response'
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Save the game
      tags:
      - Game
  /games/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scuccessfully received the desired game
          schema:
            $ref: '#/definitions/server.GameResponse'
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get a game by its ID
      tags:
      - Game
  /load/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Game name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully loaded the game
          schema:
            $ref: '#/definitions/server.Response'
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Load the game
      tags:
      - Game
  /ws:
    get:
      consumes:
      - application/json
      parameters:
      - description: Game ID
        in: query
        name: gameID
        required: true
        type: integer
      - description: Player ID
        in: query
        name: playerID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully joined the game
          schema:
            type: string
        "400":
          description: Bad request, missing fields or invalid data
          schema:
            type: string
      summary: Provided the username and game ID create a websocket connection.
      tags:
      - Game
swagger: "2.0"